Revert "Adjusted existing classes to look for moved dependencies."
This reverts commit d7361aa7236b6d8de709dc34f1ce542957b23b07.
